+++ web-client/src/components/Account/AboutProfil.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import '../App.css';
+import '../../App.css';
 
 function AboutProfil(props) {
     const styles = makeStyles({
