+++ web-client/src/components/Account/SignInForm.jsx
@@ -8,7 +8,7 @@ import * as yup from "yup";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { Link } from "react-router-dom";
-import PopUpForgotPassword from "../components/PopUpForgotPassword";
+import PopUpForgotPassword from "../PopUpForgotPassword";
 
 const styles = makeStyles({
   main: {
