+++ web-client/src/components/Account/SignInUpInfo.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
-import "../App.css";
+import "../../App.css";
 import { Link } from "react-router-dom";
 
 function SignInUpInfo(props) {
