+++ web-client/src/components/Reviews/MyReviewsOneReview.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-import '../App.css';
+import '../../App.css';
 
 function MyReviewsOneReview(props)
 {
