+++ web-client/src/components/ScientificWork/PreviewPDF.jsx
