+++ web-client/src/components/ScientificWork/ScientificWorkAuthorAnswer.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../App.css";
+import "../../App.css";
 import { makeStyles } from "@material-ui/core/styles";
 
 export default function ScientificWorkAuthorAnswer(props) {
