+++ web-client/src/components/ScientificWorkList/ScientificWorksCategories.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-import '../App.css';
+import '../../App.css';
 
-function MyReviewsCategories(props)
+function ScientificWorksCategories(props)
 {
     const styles = makeStyles({
         main:
@@ -21,10 +21,10 @@ function MyReviewsCategories(props)
 
     return(
         <div className={style.main}>
-            <Button variant='outlined' color="primary" 
+            <Button variant='contained' color="primary" 
                             className={style.btn}>{props.name}</Button>
         </div>
     )   
 }
 
-export default MyReviewsCategories; 
+export default ScientificWorksCategories;
