+++ web-client/src/components/ScientificWorkList/ScientificWorksRecentAuthors.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import '../App.css';
+import '../../App.css';
 
 function ScientificWorksRecentAuthors(props)
 {
