+++ web-client/src/pages/ScientificWorks/work-view.js
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react";
 import "../../App.css";
 import { makeStyles } from "@material-ui/core/styles";
 import { useLocation } from "react-router-dom";
-import ScienceWorkInformation from "../../components/ScienceWorkInformation";
-import VersionPanel from "../../components/VersionPanel";
+import ScienceWorkInformation from "../../components/ScientificWork/ScienceWorkInformation";
+import VersionPanel from "../../components/ScientificWork/VersionPanel";
 import axios from "axios";
 import { URL_API } from "../../Constants";
 import defaultPhoto from "../../images/empty-image.png";
